# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GoGym API",
        "description": "API REST para gerenciamento de academias, personal trainers e alunos. Permite registro de usuários (personal e aluno), autenticação via token (Sanctum), verificação de e-mail, CRUD completo de personais e alunos, e sistema de convites para vinculação entre personal e aluno.",
        "contact": {
            "name": "Gustavo",
            "email": "nao-responda@ugustavodev.com.br"
        },
        "license": {
            "name": "Proprietária",
            "url": "https://ugustavodev.com.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:84/api",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/registrar/personal": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Registrar um novo Personal Trainer",
                "description": "Cria um novo usuário com perfil de personal trainer. Um e-mail de verificação será enviado automaticamente via fila Redis.",
                "operationId": "registrarPersonal",
                "requestBody": {
                    "description": "Dados para registro do personal",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "senha",
                                    "telefone",
                                    "genero"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Carlos Silva",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos@email.com",
                                        "maxLength": 255
                                    },
                                    "senha": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123",
                                        "minLength": 4
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "(11) 99999-9999",
                                        "maxLength": 20
                                    },
                                    "genero": {
                                        "type": "string",
                                        "example": "masculino",
                                        "enum": [
                                            "masculino",
                                            "feminino",
                                            "nao_binario",
                                            "outro",
                                            "prefiro_nao_informar"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Personal registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Personal registrado com sucesso."
                                        },
                                        "dados": {
                                            "$ref": "#/components/schemas/PersonalResource"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123def456..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroValidacao"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/registrar/aluno": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Registrar um novo Aluno",
                "description": "Cria um novo usuário com perfil de aluno. Opcionalmente, pode ser vinculado a um personal trainer através de um token de convite. Um e-mail de verificação será enviado automaticamente via fila Redis.",
                "operationId": "registrarAluno",
                "requestBody": {
                    "description": "Dados para registro do aluno",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "senha",
                                    "telefone"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "João Silva",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@email.com",
                                        "maxLength": 255
                                    },
                                    "senha": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123",
                                        "minLength": 4
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "(11) 98888-8888",
                                        "maxLength": 20
                                    },
                                    "genero": {
                                        "type": "string",
                                        "example": "masculino",
                                        "nullable": true,
                                        "enum": [
                                            "masculino",
                                            "feminino",
                                            "nao_binario",
                                            "outro",
                                            "prefiro_nao_informar"
                                        ]
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1998-05-20",
                                        "nullable": true
                                    },
                                    "peso": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 75.5,
                                        "nullable": true
                                    },
                                    "altura": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.78,
                                        "nullable": true
                                    },
                                    "token_convite": {
                                        "description": "Token de convite gerado por um personal trainer",
                                        "type": "string",
                                        "example": "aB3dEfGhIjKlMnOpQrStUvWxYz...",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Aluno registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Aluno registrado com sucesso."
                                        },
                                        "dados": {
                                            "$ref": "#/components/schemas/AlunoResource"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "2|xyz789abc012..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroValidacao"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Autenticar usuário",
                "description": "Realiza login com e-mail e senha, retornando um token de acesso Bearer (Sanctum).",
                "operationId": "login",
                "requestBody": {
                    "description": "Credenciais do usuário",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "senha"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos@email.com"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "usuario": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nome": {
                                                    "type": "string",
                                                    "example": "Carlos Silva"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "carlos@email.com"
                                                },
                                                "data_verificacao_email": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "nullable": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123def456..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Os dados fornecidos são inválidos."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "As credenciais fornecidas estão incorretas."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sair": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Desconectar usuário",
                "description": "Revoga todos os tokens de acesso do usuário autenticado.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Desconectado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Desconectado com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoAutenticado"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErroValidacao": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Os dados fornecidos são inválidos."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ErroNaoAutenticado": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "PersonalResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "usuario_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "telefone": {
                        "type": "string",
                        "example": "(11) 99999-9999"
                    },
                    "genero": {
                        "type": "string",
                        "example": "masculino",
                        "enum": [
                            "masculino",
                            "feminino",
                            "nao_binario",
                            "outro",
                            "prefiro_nao_informar"
                        ]
                    },
                    "cadastrado_em": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-13 22:00:00"
                    },
                    "atualizado_em": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-13 22:00:00"
                    },
                    "deletado_em": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Insira o token retornado no login/registro. Exemplo: {token}",
                "scheme": "bearer",
                "bearerFormat": "Sanctum"
            }
        }
    },
...
}